move modules into own folder
