+++ Y.html
@@ -3,6 +3,10 @@
 
 <body>
 <section>
+<p class="h1">About this typesetting editor</p>
+</section>
+
+<section>
 <p>This typesetting editor is an application I have been developing for over half a year. The codebase was written from scratch in Python 3, and uses its own UI toolkit which I also wrote. Although still in alpha stages, the codebase can be synced from <<f class="link">https://github.com/kelvin13/shifty-octocat</f class="link">>, where it will work out of the box on any Linux installation.</p>
 
 <p>Vedit was designed to be both flexible and efficient. The end goal of the project is to include all the tools and parameters handled by “serious” typesetters, and allow editing of them in the least verbose manner. Hence, Vedit offers granular control, with a strong emphasis on <strong>additive styling</strong>.</p>
@@ -16,19 +20,19 @@
 <p style="{'align': 0}">Subtitle</p>
 
 <module:pie radius="60">
-    <module:pie:slice prop="13" color="#ff3085">
+    <module:pie:slice color="#ff3085" prop="13">
         <p>Section A</p>
     </module:pie:slice>
-    <module:pie:slice prop="22" color="#ff50a0">
+    <module:pie:slice color="#ff50a0" prop="22">
         <p>Section B</p>
     </module:pie:slice>
-    <module:pie:slice prop="69" color="#f020ff">
+    <module:pie:slice color="#f020ff" prop="69">
         <p>Section C</p>
     </module:pie:slice>
-    <module:pie:slice prop="89" color="#b050ff">
+    <module:pie:slice color="#b050ff" prop="89">
         <p>Section D</p>
     </module:pie:slice>
-    <module:pie:slice prop="113" color="#7080ff">
+    <module:pie:slice color="#7080ff" prop="113">
         <p>Section E</p>
     </module:pie:slice>
 </module:pie>
@@ -84,15 +88,11 @@
             <p>cell 4</p>
         </td>
         <td>
-            <p class="h3">PIE CHART EXAMPLE</p>
-            
-            <p style="{'align': 0}">Subtitle</p>
-            
             <module:pie radius="40">
-                <module:pie:slice prop="13" color="#ff3085">
+                <module:pie:slice color="#ff3085" prop="13">
                     <p>Section A</p>
                 </module:pie:slice>
-                <module:pie:slice prop="22" color="#ff50a0">
+                <module:pie:slice color="#ff50a0" prop="22">
                     <p>Section B</p>
                 </module:pie:slice>
             </module:pie>
@@ -126,7 +126,7 @@
         <td colspan="3">
             <p class="h2">Who doesn’t love Meredith Grey?</p>
             
-            <p style="{'leading': 200, 'align': 0, 'fontclasses': (0, [('Italic', {}), ('Regular', {'emphasis': 1})])}"><image src="meredith.png" width=200 />ab<em>c</em>d</p>
+            <p style="{'fontclasses': (0, [('Italic', {}), ('Regular', {'emphasis': 1})]), 'leading': 200, 'align': 0}"><image src="meredith.png" width=200 />ab<em>c</em>d</p>
             
             <p class="body&gray">(Taylor, evidently.)</p>
             
@@ -162,10 +162,6 @@
 
 <p></p>
 </section>
-
-<section>
-<p class="h1">About this typesetting editor</p>
-</section>
 </body>
 
 <!-- #############
@@ -183,8 +179,8 @@
                 'Superbold': {'path': '/home/kelvin/.fonts/Merriweather_UltraBold.ttf'},
                 'Superitalic': {'path': '/home/kelvin/.fonts/GaramondPremrPro-MedItCapt.otf'},
                 'h3': {'fontsize': 16, 'path': '/home/kelvin/.fonts/FiraSans-Book.otf', 'tracking': 1}},
- 'FTAGLIST': [('strong', []), ('sub', ['small']), ('emphasis', []), ('caption', []), ('big', []), ('sup', ['small']), ('link', [])],
- 'PARASTYLES': (None,
+ 'FTAGLIST': [('strong', []), ('emphasis', []), ('link', []), ('sup', ['small']), ('sub', ['small']), ('big', []), ('caption', [])],
+ 'PARASTYLES': (1,
                 [({'fontclasses': (0, [('Black', {})])}, {}),
                  ({'fontclasses': (8,
                                    [('Regular', {}),
@@ -205,10 +201,11 @@
                  ({'align': 0, 'fontclasses': (0, [('h3', {})])}, {'h3': 1}),
                  ({'fontclasses': (0, [('Gray', {})])}, {'gray': 1})]),
  'PEGS': {'Standard pegs': ('', {'sub': [0, -0.2], 'sup': [0, 0.3]})},
- 'PTAGLIST': [('gray', []), ('h2', []), ('h3', []), ('body', []), ('h1', [])],
- 'contexts': {'channels': {'c': 7, 'p': 6, 't': 0}, 'text': {'i': 4613, 'j': 4613, 'p': 6, 't': 0}},
+ 'PTAGLIST': [('body', []), ('h3', []), ('h1', []), ('gray', []), ('h2', [])],
+ 'contexts': {'channels': {'c': 7, 'p': 6, 't': 0}, 'text': {'i': 4376, 'j': 4376, 'p': 5, 't': 1}},
  'grid': [[120, 480], [120, 480]],
- 'outlines': ([(([[120, 210, False], [120, 480, False]], [[480, 210, False], [480, 480, False]]), 0),
+ 'outlines': ([(([[120, 120, False], [120, 200, True]], [[480, 120, False], [480, 200, True]]), 0)],
+              [(([[120, 210, False], [120, 480, False]], [[480, 210, False], [480, 480, False]]), 0),
                ([[[120, 120, False], [120, 480, False]], [[480, 120, False], [480, 480, False]]], 1),
                ([[[120, 120, False], [120, 460, False]], [[480, 120, False], [480, 460, False]]], 2),
                ([[[120, 120, False], [120, 480, False]], [[480, 120, False], [480, 480, False]]], 3),
@@ -216,8 +213,7 @@
                ([[[120, 120, False], [120, 480, False]], [[300, 120, False], [70, 310, False], [290, 480, False]]], 5),
                ([[[360, 120, False], [160, 310, False], [360, 480, False]], [[410, 120, False], [590, 300, False], [410, 480, False]]], 5),
                ([[[120, 120, False], [120, 480, False]], [[280, 120, False], [490, 300, True], [280, 480, False]]], 6),
-               ([[[120, 120, False], [120, 540, False]], [[480, 120, False], [480, 540, False]]], 7)],
-              [(([[120, 120, False], [120, 200, True]], [[480, 120, False], [480, 200, True]]), 0)]),
+               ([[[120, 120, False], [120, 540, False]], [[480, 120, False], [480, 540, False]]], 7)]),
  'page': {'dimensions': (600, 600), 'dual': True},
- 'view': {'H': -233, 'Hc': 460.89044189453125, 'K': -2068, 'Kc': 342.94378662109375, 'Zoom': 8, 'mode': 'text'}}
+ 'view': {'H': -58, 'Hc': 328.03839111328125, 'K': -1960, 'Kc': 235.53713989257812, 'Zoom': 9, 'mode': 'text'}}
 ############# -->
